Consolidate code.
